only addbutton will rerender after active content selection
